+++ configs/cspell/charity.txt
@@ -25,3 +25,4 @@ dotenv
 xxxs
 truetype
 svgr
+svgs
