+++ src/components/appShell/AppShell.tsx
@@ -1,7 +1,6 @@
 import { AppShell as Shell, MediaQuery } from '@mantine/core';
 
-import { Header } from '../organisms/Header';
-import { SideBar } from '../organisms/SideBar';
+import { Header, SideBar } from '../organisms';
 
 interface Props {
   children: React.ReactNode;
