--- src/components/atoms/Logo.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta } from '@storybook/react';
-
-import MainLogo from '../../assets/images/main-logo.svg';
-import { Logo } from './Logo';
-
-export default {
-  component: Logo,
-} as Meta;
-
-const image = {
-  src: MainLogo,
-  alt: 'Logo',
-};
-
-export const Default = () => <Logo logoSrc={image.src} />;
