--- src/components/atoms/Logo.tsx
@@ -1,3 +1,0 @@
-export const Logo = ({ logoSrc }: { logoSrc: string }) => {
-  return <img src={logoSrc} alt="logo" />;
-};
