+++ src/components/atoms/Logo/Logo.stories.tsx
@@ -1,0 +1,9 @@
+import type { Meta } from '@storybook/react';
+
+import { Logo } from './Logo';
+
+export default {
+  component: Logo,
+} as Meta;
+
+export const Default = () => <Logo />;
