+++ src/components/atoms/Logo/Logo.tsx
@@ -1,0 +1,5 @@
+import { MainLogoSvg } from '@camp/design';
+
+export const Logo = () => {
+  return <MainLogoSvg />;
+};
