+++ src/components/atoms/Logo/index.ts
@@ -1,0 +1,1 @@
+export * from './Logo';
