+++ src/components/atoms/NavLink/NavLink.stories.tsx
@@ -1,6 +1,6 @@
+import { DashboardIcon } from '@camp/design';
 import type { Meta, Story } from '@storybook/react';
 
-import { DashboardIcon } from '../appShell/icons/DashboardIcon';
 import { NavLink } from './NavLink';
 
 export default {
