+++ src/components/atoms/VisualElements/VisualElements.stories.tsx
@@ -1,6 +1,6 @@
+import { PeopleIcon } from '@camp/design';
 import type { Meta, Story } from '@storybook/react';
 
-import { PeopleIcon } from '../appShell/icons/PeopleIcon';
 import { VisualElements } from './VisualElements';
 
 export default {
