+++ src/components/atoms/VisualElements/VisualElements.tsx
