+++ src/components/atoms/VisualElements/index.ts
@@ -1,0 +1,1 @@
+export * from './VisualElements';
