+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export * from './appShell';
+export * from './atoms';
+export * from './molecules';
+export * from './organisms';
