+++ src/components/organisms/index.ts
@@ -1,2 +1,2 @@
 export * from './Header';
-export * from './SideBar';
+export * from './Sidebar';
