+++ src/design/icons/DashboardIcon.tsx
@@ -1,4 +1,4 @@
-import { ReactComponent as DashboardSvg } from '../../../assets/icons/Dashboard.svg';
+import { DashboardSvg } from '../svgs';
 
 export const DashboardIcon = ({ w, h }: { w: string; h: string }) => {
   return <DashboardSvg width={w} height={h} />;
