+++ src/design/icons/PackageIcon.tsx
@@ -1,4 +1,4 @@
-import { ReactComponent as PackageSvg } from '../../../assets/icons/Package.svg';
+import { PackageSvg } from '../svgs';
 
 export const PackageIcon = ({ w, h }: { w: string; h: string }) => {
   return <PackageSvg width={w} height={h} />;
