+++ src/design/icons/PeopleIcon.tsx
@@ -1,4 +1,4 @@
-import { ReactComponent as PeopleSvg } from '../../../assets/icons/people.svg';
+import { PeopleSvg } from '../svgs';
 
 export const PeopleIcon = ({ w, h }: { w: string; h: string }) => {
   return <PeopleSvg width={w} height={h} />;
