+++ src/design/icons/index.ts
@@ -1,0 +1,4 @@
+export * from './DashboardIcon';
+export * from './HomeIcon';
+export * from './PackageIcon';
+export * from './PeopleIcon';
