+++ src/design/index.ts
@@ -1,1 +1,3 @@
+export * from './icons';
+export * from './svgs';
 export * from './theme';
