+++ src/design/svgs/index.ts
@@ -1,0 +1,5 @@
+export { ReactComponent as DashboardSvg } from './Dashboard.svg';
+export { ReactComponent as HomeSvg } from './Home.svg';
+export { ReactComponent as MainLogoSvg } from './MainLogo.svg';
+export { ReactComponent as PackageSvg } from './Package.svg';
+export { ReactComponent as PeopleSvg } from './People.svg';
